Remove step to commit doc files (#218)
This PR removes the step to commit documentation files to `rapidsai/docs`.

We are overhauling that repository. All docs are now uploaded to S3 (see #214).

I didn't expect us to need to remove this step until later in `23.06`, but we're actually moving forward with this overhaul a bit sooner than I expected.